backend/requirements.txt
# Backend — prestamos.app
# Python 3.11+

# Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Supabase
supabase==2.7.4

# Settings
pydantic-settings==2.5.2

# PDF
weasyprint==62.3
jinja2==3.1.4

# HTTP client (para notificaciones)
httpx==0.27.2

# Fechas (cálculo de periodicidades en la calculadora)
python-dateutil==2.9.0

# Scheduler (job nocturno de mora)
apscheduler==3.10.4

# Utilidades
python-multipart==0.0.9
python-jose[cryptography]==3.3.0

# Dev
pytest==8.3.3
pytest-asyncio==0.24.0
httpx  # ya incluido arriba, para tests
